pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yona"
version = "0.1.0"
description = "Voice chat application for Nvidia Jetson Orin Nano with wake word detection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yona Team"}
]
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pydub>=0.25.1",
    "openwakeword>=0.5.0",
    "onnxruntime>=1.16.0",
    "faster-whisper>=0.10.0",
    "google-generativeai>=0.5.0",
    "edge-tts>=6.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
yona = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
